# src/main/resources/static/openapi.yml
openapi: 3.0.1
info:
  description: API to manage customers
  version: 1.0.0
  title: Account Service API
...
servers:
  - url: http://localhost:8080
paths:
...
  /customers/{id}/address:
    parameters:
      - $ref: '#/components/parameters/customerId'
    get:
      tags:
        - customers
      summary: Read the address of the customer
      responses:
        '200':
          description: Customer address was found
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
        '404':
          description: Customer could not be found or does not have any address assigned
    put:
      tags:
        - customers
      summary: Assign address to the customer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
        required: true
      responses:
        '201':
          description: Address was successfully assigned
        '204':
          description: Address was successfully replaced
        '400':
          description: Address is invalid
        '404':
          description: Customer could not be found or does not have any address assigned
...
components:
  schemas:
...
    Address:
      required:
        - city
        - street
        - zip
      type: object
      properties:
        street:
          minLength: 3
          maxLength: 100
          type: string
        number:
          type: string
        zip:
          pattern: '[0-9]{5}' # nur deutsche PLZ möglich
          type: string
        city:
          minLength: 1
          maxLength: 100
          type: string
...
  parameters:
    customerId:
      in: path
      name: id
      description: The id of the customer
      required: true
      schema:
        type: string
        format: uuid